Add BTC-USD as a second parallel instance; fix critical leverage bug in position sizing
Position sizing bounds were expressed as raw unit counts (min/max_units),
which only makes sense priced in a single currency's terms -- 100 units is
a sane ~$108 floor for EUR_USD but an ~$7.8M, ~780x-leveraged position for
BTC-USD (found while testing the new BTC instance). Rewrote bounds as
account-currency notional value (min_notional, max_notional_pct) so they
scale correctly across any instrument's price level.

This also required supporting fractional position sizes throughout (the
simulated broker, position sizing, and sim state) since a sane risk-sized
BTC position is routinely well under 1 whole coin -- a whole-unit floor
would make BTC trading effectively impossible at retail account sizes.
OandaClient rounds to whole units at order time, since real OANDA doesn't
accept fractional units for forex/CFDs (not a concern for BTC, which OANDA
doesn't offer).

New: BTC-USD gets its own fully independent instance (local dashboard on
:8001, cloud workflow offset 15min from EUR/USD's schedule with retry-on-
push-conflict, own dashboard page docs/btc.html) -- same pattern as the
existing local/cloud split, just one more parallel copy.

+++ live/simulated_broker.py
@@ -39,13 +39,16 @@ class SimulatedBroker:
         unrealized = units * (self._latest_price() - entry) if units != 0 else 0.0
         return {"balance": balance, "unrealizedPL": unrealized}
 
-    def get_open_units(self, instrument: str) -> int:
-        return int(get_sim_state()["sim_position_units"])
+    def get_open_units(self, instrument: str) -> float:
+        # Fractional, deliberately -- a whole-unit floor is fine for forex
+        # (1 unit of EUR is ~$1) but breaks high-priced assets like BTC-USD,
+        # where a sane risk-sized position is routinely well under 1 unit.
+        return float(get_sim_state()["sim_position_units"])
 
     def get_recent_candles(self, instrument: str, granularity: str, count: int):
         return fetch_yfinance_candles(instrument, granularity, count=count)
 
-    def place_market_order(self, instrument: str, units: int) -> dict:
+    def place_market_order(self, instrument: str, units: float) -> dict:
         state = get_sim_state()
         balance, pos, entry = state["sim_balance"], state["sim_position_units"], state["sim_entry_price"]
